Bookmarks: Change names of signals and create utils/bookmarks
